+++ apps/server/package.json
@@ -44,6 +44,7 @@
     "mongoose": "^7.0.3",
     "openai": "^3.2.1",
     "pg": "^8.8.0",
+    "pointes": "workspace:^1.0.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0",
     "socket.io": "^4.5.4",
