+++ packages/pointes/interfaces/Category.ts
@@ -1,0 +1,9 @@
+import { ProductI } from "./Product";
+
+export interface CategoryI {
+  id: string;
+  description: string;
+  parent: ProductI;
+  children: ProductI[];
+  name: string;
+}
