+++ packages/pointes/interfaces/Item.ts
@@ -1,0 +1,11 @@
+export interface ItemI {
+  configName: string;
+  altTxt: string;
+  image?: string;
+  sku: string;
+  upc?: string;
+  price: number;
+  status: number;
+  stock: number;
+  product?: any;
+}
