+++ packages/pointes/package.json
@@ -1,15 +1,12 @@
 {
   "name": "pointes",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "mongoose": "^7.0.3"
-  }
+  "license": "ISC"
 }
