+++ packages/pointes/types/index.ts
@@ -1,14 +1,30 @@
-import { Types } from 'mongoose';
 
-export type CategoryCreateType = {
-  name: string;
-  description: string;
-  parent?: string | Types.ObjectId;
-};
-
-export type SupplierCreateType = {
-  name: string;
-  description: string;
-  logo: string;
-  location?: string;
-};
+export type SupabaseJWTPayload = {
+  aud: string;
+  exp: number;
+  sub: string;
+  email: string;
+  phone: string;
+  app_metadata: {
+    provider: string;
+    providers: string[];
+  };
+  user_metadata: {
+    avatar_url: string;
+    email: string;
+    email_verified: boolean;
+    full_name: string;
+    iss: string;
+    name: string;
+    picture: string;
+    provider_id: string;
+    sub: string;
+  };
+  role: string;
+  aal: string;
+  amr: {
+    method: string;
+    timestamp: number;
+  }[];
+  session_id: string;
+}
