+++ packages/pointes/types/request.ts
@@ -1,0 +1,12 @@
+export type CategoryCreateType = {
+  name: string;
+  description: string;
+  parent?: string;
+};
+
+export type SupplierCreateType = {
+  name: string;
+  description: string;
+  logo: string;
+  location?: string;
+};
